requirements.txt
aiogram~=3.0.0rc1
python-dotenv~=1.0.0
SQLAlchemy~=2.0.19
alembic~=1.11.2
asyncpg~=0.28.0